+++ Sources/Paywall/Storage/Storage.swift
@@ -174,4 +174,26 @@ class Storage {
       forType: Transactions.self
     )
   }
+
+  func saveLastPaywallView() {
+    cache.write(
+      Date(),
+      forType: LastPaywallView.self
+    )
+  }
+
+  func getLastPaywallView() -> LastPaywallView.Value? {
+    return cache.read(LastPaywallView.self)
+  }
+
+  func incrementTotalPaywallViews() {
+    cache.write(
+      (getTotalPaywallViews() ?? 0) + 1,
+      forType: TotalPaywallViews.self
+    )
+  }
+
+  func getTotalPaywallViews() -> TotalPaywallViews.Value? {
+    return cache.read(TotalPaywallViews.self)
+  }
 }